Removed old page remove and further reduced entity repo
